Transfer getSubdomainAndPath to domain_parsing.ts & create tests
