+++ portal/src/helpers/domain_parsing.test.ts
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { describe, expect, test } from 'vitest'
-import { getDomain } from './domain_parsing'
+import { getDomain, getSubdomainAndPath } from './domain_parsing'
 
-const testCases: [string, string][] = [
+const getDomainTestCases: [string, string][] = [
     ['https://example.com', 'example.com'],
     ['https://suinsname.localhost:8080', 'localhost'],
     ['https://subname.suinsname.localhost:8080', 'localhost'],
@@ -19,10 +19,26 @@ const testCases: [string, string][] = [
 ]
 
 describe('getDomain', () => {
-    testCases.forEach(([input, expected]) => {
+    getDomainTestCases.forEach(([input, expected]) => {
         test(`${input} -> ${expected}`, () => {
             const domain = getDomain(input)
                 expect(domain).toEqual(expected)
         })
     })
+})
+
+const getSubdomainAndPathTestCases: [string, string, string][] = [
+  ['https://flatland.walrus.site/', 'flatland', '/index.html'],
+]
+
+describe('getSubdomainAndPath', () => {
+  getSubdomainAndPathTestCases.forEach(
+    ([input, expectedSubdomain, expectedPath]) => {
+    test(`${input} -> subdomain: ${expectedSubdomain}, path: ${expectedPath}`,
+      () => {
+          console.log(
+              'getSubdomainAndPath',
+              getSubdomainAndPath(new URL('https://subname.flatland.walrus.site/')))
+    });
+  });
 })
